archive/demos/02-agent-service/02-knowledge-tools/agent-knowledge-tools-py/pyproject.toml
[project]
name = "agent-knowledge-tools-py"
version = "0.1.0"
description = "AI Agent demo with function calling and knowledge tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "azure-identity>=1.16.0",
    "azure-ai-projects>=2.0.0b2",
    "azure-ai-agents>=1.2.0b6",
    "requests>=2.31.0"
]